% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Dataset title.}
\format{
A \code{\link[tibble]{tibble}} with 224 rows and 29 variables:
  \describe{
    \item{index}{...}
    \item{subject}{Subject ID.}
 }
}
\source{
...
}
\usage{
dataset
}
\description{
A tibble ...
}
\keyword{datasets}
