% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emotion}
\alias{emotion}
\title{Word emotional valence}
\format{
A \code{\link[tibble]{tibble}} with 1000 rows and 2 variables:
  \describe{
    \item{word}{Word.}
    \item{valency_z}{Emotional valency of word (z-scores).}
  }
}
\usage{
emotion
}
\description{
This tibble contains data from Warriner 2013.
}
\keyword{datasets}
