% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incomplete}
\alias{incomplete}
\title{Incomplete neutralisation in German}
\format{
A \code{\link[tibble]{tibble}} with 6144 rows and 8 variables:
  \describe{
    \item{order}{Trial number.}
    \item{speaker_voice}{Speaker voice ID.}
    \item{item_pair}{ID of the lexical pair (voiceless/voiced).}
    \item{RT}{Reaction time.}
    \item{correct}{Whether the listener has correctly categorised the word.}
    \item{correct_voicing}{The underlying voicing category of the word.}
    \item{listener}{The listener ID.}
    \item{repetitiontype}{Whether the word was heard for the first time or subsequent times.}
  }
}
\usage{
incomplete
}
\description{
This tibble contains data from Roettger et al. 2014.
}
\keyword{datasets}
