% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnB4SS-package.R
\docType{package}
\name{learnB4SS-package}
\alias{learnB4SS}
\alias{learnB4SS-package}
\title{learnB4SS: Learning materials for the learnB4SS workshop}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It provides the learning materials from the workshop 'Learn Bayesian Analysis for the Speech Sciences'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/learnB4SS/learnB4SS}
  \item \url{https://learnb4ss.github.io/learnB4SS/}
  \item Report bugs at \url{https://github.com/learnB4SS/learnB4SS/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID})

Authors:
\itemize{
  \item Joseph V. Casillas (\href{https://orcid.org/0000-0001-8735-9910}{ORCID})
  \item Timo Roettger (\href{https://orcid.org/0000-0003-1400-2739}{ORCID})
}

}
\keyword{internal}
