% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_exercise}
\alias{open_exercise}
\title{Open exercise file.}
\usage{
open_exercise(session)
}
\arguments{
\item{session}{Session number as a number.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Call this function with the session number as argument and the exercise
file will be copied in the working directory and opened in the editor.
}
\details{
The following sessions have exercises:

- Session 03.

- Session 06.

- Session 08.

You can also open a fully worked-out analysis with `open_exercise("full")`
}
\examples{
\dontrun{
open_exercise(3)

open_exercise("full")
}
# The exercise of Session 03 will open for editing.
}
