% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_slides}
\alias{open_slides}
\title{Open slides in browser.}
\usage{
open_slides(session)
}
\arguments{
\item{session}{Session number as a number.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Call this function with the session number as argument and the slides of that
session will open in your browser (an internet connection is required).
}
\examples{
\dontrun{
open_slides(0)
}
}
