% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Vowel duration and consonant voicing in Italian}
\format{
A \code{\link[tibble]{tibble}} with 6144 rows and 8 variables:
  \describe{
    \item{item}{Trial number.}
    \item{speaker}{Speaker ID.}
    \item{word}{Target word.}
    \item{v1_duration}{Duration of V1 in milliseconds.}
    \item{c2_voicing}{Voicing of C2 (voiceless, voiced).}
    \item{vowel}{Target vowel.}
    \item{c2_place}{Place of articulation of C2.}
    \item{speech_rate}{Number of syllables per second.}
    \item{speech_rate_c}{Centred speech rate.}
  }
}
\source{
Coretta, Stefano. 2020. Vowel duration and consonant voicing: A production study.
    PhD Thesis. University of Manchester. https://doi.org/10.17605/OSF.IO/W92ME
}
\usage{
vowels
}
\description{
This tibble contains data from Coretta 2020.
}
\keyword{datasets}
