% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emotion}
\alias{emotion}
\title{Word emotional valence}
\format{
A \code{\link[tibble]{tibble}} with 1000 rows and 2 variables:
  \describe{
    \item{word}{Word.}
    \item{valency_z}{Emotional valency of word (z-scores).}
  }
}
\source{
Warriner, A. B., Kuperman, V., & Brysbaert, M. 2013. Norms of
    valence, arousal, and dominance for 13,915 English lemmas. Behavior
    research methods, 45(4), 1191–1207.
    https://doi.org/10.3758/s13428-012-0314-x
}
\usage{
emotion
}
\description{
This tibble contains data from Warriner 2013.
}
\keyword{datasets}
