% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_exercise}
\alias{open_exercise}
\title{Open exercise file.}
\usage{
open_exercise(session)
}
\arguments{
\item{session}{Session number as a number.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Call this function with the session number as argument and the exercise
file will be copied in the working directory and opened in the editor.
}
\examples{
\dontrun{
open_exercise(1)
}
# The exercise of Session 01 will open for editing.
}
